+++ packages/d-pi/CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Fixed
 
 - Fixed npm publish E422 due to missing `repository.url` (will be in v0.6.0-alpha.3)
+- Fixed pre-existing 19+1 vitest false positives via `@sheason/*` alias in vitest config